teste unitário e teste de integração

+++ sifatecommerce/src/main/java/br/com/sifatecommerce/service/CategoriaService.java
@@ -1,13 +1,18 @@
 package br.com.sifatecommerce.service;
 
+import java.util.List;
 import java.util.Optional;
 
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.web.server.ResponseStatusException;
 
 import br.com.sifatecommerce.entity.Categoria;
+import br.com.sifatecommerce.entity.Produto;
 import br.com.sifatecommerce.repository.CategoriaRepository;
+import br.com.sifatecommerce.repository.ProdutoRepository;
 import lombok.RequiredArgsConstructor;
 
 @Service
@@ -15,6 +20,7 @@ import lombok.RequiredArgsConstructor;
 public class CategoriaService {
 
 	private final CategoriaRepository  categoriaRepository;
+	private final ProdutoRepository produtoRepository;
 	
 	public Categoria salvarCategoria(Categoria categoria) {
 		return categoriaRepository.save(categoria);
@@ -24,11 +30,15 @@ public class CategoriaService {
 		return categoriaRepository.findById(id);
 	}
 
-	public void deletarCategoria(Long id) {
+	public void  deletarCategoria(Long id) {
+		List<Produto> produtos = produtoRepository.findByCategoriaId(id);
+		if(!produtos.isEmpty()) {
+			throw new  ResponseStatusException(HttpStatus.BAD_REQUEST, "Há produtos vinculado a categoria"); 
+		}
 	    categoriaRepository.findById(id).map(categoria -> {
 	    	categoriaRepository.delete(categoria);
 	        return categoria;
-	    }).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Categoria não encontrada!")); // Lança exceção se não encontrado
+	    }).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Categoria não encontrada!")); 
 	}
 
 	
@@ -40,5 +50,9 @@ public class CategoriaService {
 	        }).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Categoria não encontrada!"));
 	}
 
+	public Page<Categoria> procurarPorAtributos(String query, int page, int pageSize) {
+        PageRequest pageRequest = PageRequest.of(page, pageSize);
+        return categoriaRepository.findByNomeContainingIgnoreCase(query, pageRequest);
+    }
 	
 }
